More refactoring, set up other pages
