+++ myApp/TODO
@@ -1,3 +1,4 @@
+G shape no min7
 C shape min9
 D root
