+++ myApp/src/components/chords.ts
@@ -1,5 +1,5 @@
 import { colors, noteCircle, fretBoxes, strings, generateFretDegrees } from "./frets"
-import { Quality, Root, Chord } from './types'
+import { Shape, Quality, Root, Chord } from './types'
 
 function generateChord(notes: number[][], root: Root) : Chord[] {
   return notes.map((note) => ({
@@ -15,6 +15,39 @@ function generateChord(notes: number[][], root: Root) : Chord[] {
       }
     })
   )
+}
+
+function getChordDescription(quality: Quality, extensions: string[], alterations: string[]) {
+  const empty = "Use the below dropdowns."
+    let strQuality = ""
+    const extAndAlt = []
+
+    if (quality.includes("7") || quality.includes("9")) {
+      if (quality.includes("maj")) {
+        strQuality = "Major"
+      } else if (quality.includes("min")) {
+        strQuality = "Minor"
+      } else {
+        strQuality = "Dominant"
+      }
+      if (!alterations.includes("maj7")) extAndAlt.push("7")
+      if (alterations.includes("b5")) extAndAlt.push("b5")
+      if (alterations.includes("#5")) extAndAlt.push("#5")
+      if (quality.includes("9") || extensions.includes("9")) {
+        if (alterations.includes("b9")) extAndAlt.push("b9")
+        else if (alterations.includes("#9")) extAndAlt.push("#9")
+        else extAndAlt.push("9")
+      }
+      if (alterations.includes("#11")) extAndAlt.push("#11")
+      if (extensions.includes("13")) {
+        if (alterations.includes("b13")) extAndAlt.push("b13")
+        else extAndAlt.push("13")
+      }
+    }
+    return strQuality !== "" ? extAndAlt.length !== 0 ? 
+      `${strQuality} ${extAndAlt.reduce((total, current, i) => (total + (i<extAndAlt.length ? ", ": "") + current))}` 
+      : strQuality 
+      : empty
 }
 
 const chordE7 = (quality: Quality, extensions: string[], alterations: string[]) => {
@@ -113,12 +146,37 @@ const chordA7 = (quality: Quality, extensions: string[], alterations: string[]) 
   return chord
 }
 
+function getChord(root: Root, shape: Shape, quality: Quality, extensions: string[], alterations: string[]) {
+  let chord: number[][] = [[]]
+  if (root === "E") {
+    if (shape === "E" && quality.includes("7")) {
+      chord = chordE7(quality, extensions, alterations)
+    } else if (shape === "G" && quality.includes("7")) {
+      chord = chordG7(quality, extensions, alterations)
+    } else if (quality.includes("9")) {
+      chord = chordE9(quality, extensions, alterations)
+    } else {
+      chord = [[5,3]]
+    }
+  }
+  if (root === "A") {
+    if (shape === "C" && quality.includes("7")) {
+      chord = chordC7(quality, extensions, alterations)
+    } else if (shape === "A" && quality.includes("7")) {
+      chord = chordA7(quality, extensions, alterations)
+    } else if (quality.includes("9")) {
+      chord = chordC9(quality, extensions, alterations)
+    } else {
+      chord = [[4,3]]
+    }
+  }
+  if (root === "D") {
+    chord = [[3,3]]
+  }
+  return generateChord(chord, root)
+}
+
 export {
-  generateChord,
-  chordE7,
-  chordE9,
-  chordG7,
-  chordA7,
-  chordC7,
-  chordC9
+  getChordDescription,
+  getChord
 }
