+++ myApp/src/pages/Concepts.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/react';
-import './Tab2.css';
+import './Concepts.css';
 
-const Tab2: React.FC = () => {
+const Concepts: React.FC = () => {
   return (
     <IonPage>
       <IonHeader>
         <IonToolbar>
-          <IonTitle>Tab 2</IonTitle>
+          <IonTitle>Concepts</IonTitle>
         </IonToolbar>
       </IonHeader>
       <IonContent fullscreen>
         <IonHeader collapse="condense">
           <IonToolbar>
-            <IonTitle size="large">Tab 2</IonTitle>
+            <IonTitle size="large">Concepts</IonTitle>
           </IonToolbar>
         </IonHeader>
       </IonContent>
@@ -21,4 +21,4 @@ const Tab2: React.FC = () => {
   );
 };
 
-export default Tab2;
+export default Concepts;
